requirements.txt
# Python Dependencies for DigitalFTE

# Core & Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
pyyaml>=6.0

# Google API (Gmail)
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.100.0

# Web Automation (WhatsApp)
playwright>=1.40.0

# File System Monitoring
watchdog>=3.0.0

# Xero Integration
xero-python>=4.0.0

# Social Media (Twitter)
tweepy>=4.14.0

# Data Processing
pandas>=2.1.0
openpyxl>=3.1.0

# HTTP & API
requests>=2.31.0
aiohttp>=3.9.0

# Logging & Monitoring
python-json-logger>=2.0.7
coloredlogs>=15.0.1

# Scheduling
APScheduler>=3.10.0

# Database (optional - for future)
sqlalchemy>=2.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Code Quality
black>=23.12.0
flake8>=6.1.0
isort>=5.13.0
mypy>=1.7.0

# Development
ipython>=8.18.0
jupyter>=1.0.0
